+++ src/classes/DataBoard.ts
@@ -4,35 +4,68 @@ import { dict, HOUR } from "../platform/helpers";
 
 /*
  * decaffeinate suggestions:
- * DS102: Remove unnecessary code created because of implicit returns
  * DS104: Avoid inline assignments
- * DS206: Consider reworking classes to avoid initClass
  * DS207: Consider shorter variations of null checks
  * Full docs: https://github.com/decaffeinate/decaffeinate/blob/main/docs/suggestions.md
  */
-export default class DataBoard {
-  static initClass() {
-    this.keys = ['hiddenThreads', 'hiddenPosts', 'lastReadPosts', 'yourPosts', 'watchedThreads', 'watcherLastModified', 'customTitles'];
-  }
 
-  constructor(key, sync, dontClean) {
+type DataBoardData = {
+  [site: string]: {
+    boards: {
+      [threadId: string]: number;
+    };
+    lastChecked?: number;
+  };
+} & { version?: number };
+
+interface PostInfo {
+  /** Defaults to g.SITE.ID */
+  siteID?: string,
+  boardID: string,
+  threadID ?: string | number,
+  postID ?: string | number
+};
+
+/**
+ * This class handles data related to specific threads or posts. This data is automatically cleaned up when the thread
+ * ages out.
+ * TODO At this moment, .get and .set aren't fully typed yet.
+ */
+export default class DataBoard {
+  static keys = [
+    'hiddenThreads',
+    'hiddenPosts',
+    'hiddenPosterIds',
+    'lastReadPosts',
+    'yourPosts',
+    'watchedThreads',
+    'watcherLastModified',
+    'customTitles',
+  ] as const;
+
+  declare data: DataBoardData;
+  declare changes: any[];
+  declare key: (typeof DataBoard.keys)[number];
+  declare sync?: () => void;
+
+  constructor(key: (typeof DataBoard.keys)[number], sync?: () => void, dontClean = false) {
     this.changes = [];
     this.onSync = this.onSync.bind(this);
     this.key = key;
     this.initData(Conf[this.key]);
     $.sync(this.key, this.onSync);
-    if (!dontClean) { this.clean(); }
-    if (!sync) { return; }
+    if (!dontClean) this.clean();
+    if (!sync) return;
     // Chrome also fires the onChanged callback on the current tab,
     // so we only start syncing when we're ready.
     var init = () => {
       $.off(d, '4chanXInitFinished', init);
-      return this.sync = sync;
+      this.sync = sync;
     };
     $.on(d, '4chanXInitFinished', init);
   }
 
-  initData(data) {
+  initData(data: DataBoardData) {
     let boards;
     this.data = data;
     if (this.data.boards) {
@@ -45,10 +78,10 @@ export default class DataBoard {
     return this.data[g.SITE.ID] || (this.data[g.SITE.ID] = { boards: dict() });
   }
 
-  save(change, cb) {
+  save(change: () => void, cb?: () => void) {
     change();
     this.changes.push(change);
-    return $.get(this.key, { boards: dict() }, items => {
+    return $.get(this.key, { boards: dict() }, (items: DataBoardData) => {
       if (!this.changes.length) { return; }
       const needSync = ((items[this.key].version || 0) > (this.data.version || 0));
       if (needSync) {
@@ -65,7 +98,7 @@ export default class DataBoard {
   }
 
   forceSync(cb) {
-    return $.get(this.key, { boards: dict() }, items => {
+    return $.get(this.key, { boards: dict() }, (items: DataBoardData) => {
       if ((items[this.key].version || 0) > (this.data.version || 0)) {
         this.initData(items[this.key]);
         for (var change of this.changes) { change(); }
@@ -78,56 +111,59 @@ export default class DataBoard {
   delete({siteID, boardID, threadID, postID}, cb) {
     if (!siteID) { siteID = g.SITE.ID; }
     if (!this.data[siteID]) { return; }
-    return this.save(() => {
+    this.save(() => {
       if (postID) {
         if (!this.data[siteID].boards[boardID]?.[threadID]) { return; }
         delete this.data[siteID].boards[boardID][threadID][postID];
-        return this.deleteIfEmpty({siteID, boardID, threadID});
+        this.deleteIfEmpty({siteID, boardID, threadID});
       } else if (threadID) {
         if (!this.data[siteID].boards[boardID]) { return; }
         delete this.data[siteID].boards[boardID][threadID];
-        return this.deleteIfEmpty({siteID, boardID});
+        this.deleteIfEmpty({siteID, boardID});
       } else {
-        return delete this.data[siteID].boards[boardID];
+        delete this.data[siteID].boards[boardID];
       }
     }
     , cb);
   }
 
-  deleteIfEmpty({siteID, boardID, threadID}) {
+  deleteIfEmpty({ siteID, boardID, threadID }: { siteID: string, boardID: string, threadID?: string | number }) {
     if (!this.data[siteID]) { return; }
     if (threadID) {
       if (!Object.keys(this.data[siteID].boards[boardID][threadID]).length) {
         delete this.data[siteID].boards[boardID][threadID];
-        return this.deleteIfEmpty({siteID, boardID});
+        this.deleteIfEmpty({siteID, boardID});
       }
     } else if (!Object.keys(this.data[siteID].boards[boardID]).length) {
-      return delete this.data[siteID].boards[boardID];
+      delete this.data[siteID].boards[boardID];
     }
   }
 
-  set(data, cb) {
-    return this.save(() => {
-      return this.setUnsafe(data);
-    }
-    , cb);
+  set(data: Parameters<DataBoard['setUnsafe']>[0], cb?: () => void) {
+    this.save(() => {
+      this.setUnsafe(data);
+    }, cb);
   }
 
-  setUnsafe({siteID, boardID, threadID, postID, val}) {
+  setUnsafe({ siteID, boardID, threadID, postID, val }: PostInfo & { val?: any }) {
     if (!siteID) { siteID = g.SITE.ID; }
-    if (!this.data[siteID]) { this.data[siteID] = { boards: dict() }; }
+    if (!this.data[siteID]) this.data[siteID] = { boards: dict() };
+    const boards = this.data[siteID].boards;
     if (postID !== undefined) {
       let base;
-      return (((base = this.data[siteID].boards[boardID] || (this.data[siteID].boards[boardID] = dict())))[threadID] || (base[threadID] = dict()))[postID] = val;
+      (((base = boards[boardID] || (boards[boardID] = dict())))[threadID] || (base[threadID] = dict()))[postID] = val;
     } else if (threadID !== undefined) {
-      return (this.data[siteID].boards[boardID] || (this.data[siteID].boards[boardID] = dict()))[threadID] = val;
+      (boards[boardID] || (boards[boardID] = dict()))[threadID] = val;
     } else {
-      return this.data[siteID].boards[boardID] = val;
+      boards[boardID] = val;
     }
   }
 
-  extend({siteID, boardID, threadID, postID, val}, cb) {
-    return this.save(() => {
+  extend(
+    { siteID, boardID, threadID, postID, val }: PostInfo & { val?: any },
+    cb?: () => void,
+  ) {
+    this.save(() => {
       const oldVal = this.get({ siteID, boardID, threadID, postID, defaultValue: dict() });
       for (var key in val) {
         var subVal = val[key];
@@ -137,18 +173,17 @@ export default class DataBoard {
           oldVal[key] = subVal;
         }
       }
-      return this.setUnsafe({siteID, boardID, threadID, postID, val: oldVal});
-    }
-    , cb);
+      this.setUnsafe({siteID, boardID, threadID, postID, val: oldVal});
+    }, cb);
   }
 
   setLastChecked(key='lastChecked') {
-    return this.save(() => {
-      return this.data[key] = Date.now();
+    this.save(() => {
+      this.data[key] = Date.now();
     });
   }
 
-  get({siteID, boardID, threadID, postID, defaultValue}) {
+  get({ siteID, boardID, threadID, postID, defaultValue }: PostInfo & { defaultValue?: any }) {
     let board, val;
     if (!siteID) { siteID = g.SITE.ID; }
     if (board = this.data[siteID]?.boards[boardID]) {
@@ -156,7 +191,6 @@ export default class DataBoard {
       if (threadID == null) {
         if (postID != null) {
           for (thread = 0; thread < board.length; thread++) {
-            var ID = board[thread];
             if (postID in thread) {
               val = thread[postID];
               break;
@@ -166,10 +200,7 @@ export default class DataBoard {
           val = board;
         }
       } else if (thread = board[threadID]) {
-        val = (postID != null) ?
-          thread[postID]
-        :
-          thread;
+        val = (postID != null) ? thread[postID] : thread;
       }
     }
     return val || defaultValue;
@@ -179,7 +210,6 @@ export default class DataBoard {
     let boardID, middle;
     const siteID = g.SITE.ID;
     for (boardID in this.data[siteID].boards) {
-      var val = this.data[siteID].boards[boardID];
       this.deleteIfEmpty({siteID, boardID});
     }
     const now = Date.now();
@@ -196,22 +226,22 @@ export default class DataBoard {
     const siteID = g.SITE.ID;
     const threadsList = g.SITE.urls.threadsListJSON?.({siteID, boardID});
     if (!threadsList) { return; }
-    return $.cache(threadsList, function() {
+    $.cache(threadsList, function() {
       if (this.status !== 200) { return; }
       const archiveList = g.SITE.urls.archiveListJSON?.({siteID, boardID});
-      if (!archiveList) { return that.ajaxCleanParse(boardID, this.response); }
+      if (!archiveList) return that.ajaxCleanParse(boardID, this.response);
       const response1 = this.response;
-      return $.cache(archiveList, function() {
+      $.cache(archiveList, function() {
         if ((this.status !== 200) && (!!g.SITE.archivedBoardsKnown || (this.status !== 404))) { return; }
-        return that.ajaxCleanParse(boardID, response1, this.response);
+        that.ajaxCleanParse(boardID, response1, this.response);
       });
     });
   }
 
-  ajaxCleanParse(boardID, response1, response2) {
+  ajaxCleanParse(boardID: string, response1: any, response2?: any) {
     let board, ID;
     const siteID = g.SITE.ID;
-    if (!(board = this.data[siteID].boards[boardID])) { return; }
+    if (!(board = this.data[siteID].boards[boardID])) return;
     const threads = dict();
     if (response1) {
       for (var page of response1) {
@@ -223,18 +253,17 @@ export default class DataBoard {
     }
     if (response2) {
       for (ID of response2) {
-        if (ID in board) { threads[ID] = board[ID]; }
+        if (ID in board) threads[ID] = board[ID];
       }
     }
     this.data[siteID].boards[boardID] = threads;
     this.deleteIfEmpty({siteID, boardID});
-    return $.set(this.key, this.data);
+    $.set(this.key, this.data);
   }
 
   onSync(data) {
     if ((data.version || 0) <= (this.data.version || 0)) { return; }
     this.initData(data);
-    return this.sync?.();
+    this.sync?.();
   }
 }
-DataBoard.initClass();